movement basic is working

+++ crates/game/src/character/player/jjrs.rs
@@ -8,6 +8,6 @@ use serde::{Deserialize, Serialize};
 
 #[repr(C)]
 #[derive(Copy, Clone, Debug, PartialEq, Eq, Default, Serialize, Deserialize)]
-pub struct BoxInput(pub u8);
+pub struct BoxInput(pub u16);
 
 pub type BoxConfig = GgrsConfig<BoxInput>;
